+++ gitman/__init__.py
@@ -11,7 +11,7 @@ NAME = "Git Dependency Manager"
 VERSION = __project__ + ' v' + __version__
 DESCRIPTION = "A language-agnostic dependency manager using Git."
 
-PYTHON_VERSION = 3, 4
+PYTHON_VERSION = 3, 5
 
 if sys.version_info < PYTHON_VERSION:  # pragma: no cover (manual test)
     sys.exit("Python {}.{}+ is required.".format(*PYTHON_VERSION))