+++ gitman/shell.py
@@ -14,7 +14,14 @@ log = logging.getLogger(__name__)
 
 
 def call(name, *args, _show=True, _ignore=False):
-    """Call a shell program with arguments."""
+    """Call a shell program with arguments.
+
+    :param name: name of program to call
+    :param args: list of command-line arguments
+    :param _show: display the call on stdout
+    :param _ignore: ignore non-zero return codes
+
+    """
     program = CMD_PREFIX + ' '.join([name, *args])
     if _show:
         common.show(program)
@@ -22,14 +29,13 @@ def call(name, *args, _show=True, _ignore=False):
         log.debug(program)
 
     if name == 'cd':
-        assert len(args) == 1, "'cd' takes a single argument"
-        return os.chdir(args[0])
+        return os.chdir(args[0])  # 'cd' has no effect in a subprocess
 
     command = subprocess.run(
-        [name, *args],
+        [name, *args], universal_newlines=True,
         stdout=subprocess.PIPE, stderr=subprocess.STDOUT,
-        universal_newlines=True,
     )
+
     for line in command.stdout.splitlines():
         log.debug(OUT_PREFIX + line.strip())
 
@@ -40,14 +46,14 @@ def call(name, *args, _show=True, _ignore=False):
         log.debug("Ignored error from call to '%s'", program)
 
     else:
-        msg = (
+        message = (
             "An external program call failed." + "\n\n"
             "In worikng directory: " + os.getcwd() + "\n\n"
             "Execution produced a non-zero return code:" + "\n\n" +
             program + "\n" +
             command.stdout
         )
-        raise ShellError(msg)
+        raise ShellError(message)
 
 
 def mkdir(path):
@@ -55,7 +61,6 @@ def mkdir(path):
 
 
 def cd(path, _show=True):
-
     call('cd', path, _show=_show)